backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
authlib
passlib[bcrypt]
pydantic[email]